functions/requirements.txt
firebase_functions~=0.1.0
firebase-admin==6.4.0
google-cloud-firestore
functions-framework